代码整理 pod 修改 首页视频样式修改

+++ scholars/Other/DataUtil/ORDateUtil.m
@@ -16,7 +16,7 @@
 {
     NSString *dateString = nil;
     
-    time_t intervalInSecond = aTimeInterval;
+    time_t intervalInSecond = (long)aTimeInterval;
     time_t ltime;
     time(&ltime);
